requirements.txt
# PnL Research - Python Dependencies

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0

# Environment Variables
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Async Support
asyncio-throttle>=1.0.0

# Google Gemini API (optional, for AI features)
google-genai>=0.3.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
colorlog>=6.7.0
